+++ changelog.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [2.0.2] - [Unreleased]
+
+### Added
+- Now we can pass custom callback `onNavClose` to be invoked when Primary or Secondary Navigation Menu is closed
+- Added `enzyme-adapter-react-16@1.0.2` to make test cases compatible with `react 16`
+
+### Changed
+- Updated `enzyme` to `3.1.0` to make test cases compatible with `react 16`.
+
+### Fixed
+- Some type issues.
+- `react-select` import in `DropDownInputTemplate`.
+- Fixed `onChange` handler in `DropDownInputTemplate`.
+
 ## [2.0.1] - 2017-10-31
 
 ### Added